update: prevented to remove the default image

+++ backend/src/utils/removeFile.ts
@@ -2,7 +2,8 @@ import fs from 'fs'
 import path from 'path'
 
 export async function removeFile(file: string){
-    const filePath = path.join(__dirname, '..', 'uploads', file)
+    if(file=="default.jpg") return
+    const filePath = path.join(__dirname, '..','..', 'uploads', file)
     fs.unlink(filePath, (error)=>{
         if(error) {
             console.log(error)
